+++ CHANGELOG.md
@@ -36,6 +36,8 @@
   no longer requires the option. We'll automatically upgrade to 23.0 if we need to.
 - Fix documentation for Vault List, Get, Post et al commands
 - Fix regression with echo on/off (but not no-args or expr) echoing itself.
+- Fix for documentation missing for Is Defined, Is Empty, Is Constant and their
+  negated versions.
 
 ### Security